requirements.txt
asgiref==3.4.1
boto3==1.18.23
botocore==1.21.23
certifi==2021.5.30
Django==3.2.6
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.3.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
